+++ pypsa/components/store.py
@@ -7,10 +7,8 @@ from __future__ import annotations
 
 import logging
 import re
-import warnings
 from typing import TYPE_CHECKING, Any
 
-from pypsa._options import options
 from pypsa.deprecations import COMPONENT_ALIAS_DICT
 
 if TYPE_CHECKING:
@@ -65,21 +63,21 @@ class ComponentsStore(dict):
         >>> n.components
         PyPSA Components Store
         ======================
-        - 3 'SubNetwork' Components
         - 9 'Bus' Components
         - 6 'Carrier' Components
-        - 1 'GlobalConstraint' Components
+        - 6 'Generator' Components
+        - 6 'Load' Components
+        - 4 'Link' Components
+        - 0 'Store' Components
+        - 0 'StorageUnit' Components
         - 7 'Line' Components
         - 36 'LineType' Components
         - 0 'Transformer' Components
         - 14 'TransformerType' Components
-        - 4 'Link' Components
-        - 6 'Load' Components
-        - 6 'Generator' Components
-        - 0 'StorageUnit' Components
-        - 0 'Store' Components
         - 0 'ShuntImpedance' Components
+        - 1 'GlobalConstraint' Components
         - 0 'Shape' Components
+        - 3 'SubNetwork' Components
 
         """
         return "PyPSA Components Store\n======================\n- " + "\n- ".join(
@@ -103,21 +101,21 @@ class ComponentsStore(dict):
         >>> n.components
         PyPSA Components Store
         ======================
-        - 3 'SubNetwork' Components
         - 9 'Bus' Components
         - 6 'Carrier' Components
-        - 1 'GlobalConstraint' Components
+        - 6 'Generator' Components
+        - 6 'Load' Components
+        - 4 'Link' Components
+        - 0 'Store' Components
+        - 0 'StorageUnit' Components
         - 7 'Line' Components
         - 36 'LineType' Components
         - 0 'Transformer' Components
         - 14 'TransformerType' Components
-        - 4 'Link' Components
-        - 6 'Load' Components
-        - 6 'Generator' Components
-        - 0 'StorageUnit' Components
-        - 0 'Store' Components
         - 0 'ShuntImpedance' Components
+        - 1 'GlobalConstraint' Components
         - 0 'Shape' Components
+        - 3 'SubNetwork' Components
         >>> n.components["generators"]
         'Generator' Components
         ----------------------
@@ -169,15 +167,6 @@ class ComponentsStore(dict):
 
     def __iter__(self) -> Any:
         """Value iterator over components in store."""
-        if options.get_option("warnings.components_store_iter"):
-            warnings.warn(
-                "Iterating over `n.components` yields the values instead of keys from "
-                "v0.33.0. This behavior might be breaking. Use `n.components.keys()` "
-                "to iterate over the keys. To suppress this warning set "
-                "`pypsa.options.warnings.components_store_iter = False`.",
-                DeprecationWarning,
-                stacklevel=2,
-            )
         return iter(self.values())
 
     def __contains__(self, item: Any) -> bool: