+++ pypsa/optimization/expressions.py
@@ -16,7 +16,6 @@ from pypsa.common import pass_none_if_keyerror
 from pypsa.descriptors import nominal_attrs
 from pypsa.statistics import (
     get_transmission_branches,
-    get_weightings,
     port_efficiency,
 )
 from pypsa.statistics.abstract import AbstractStatisticsAccessor
@@ -33,6 +32,7 @@ USE_EMPTY_PROPERTY = version.parse(ln.__version__) >= version.parse("0.5.1")
 
 def check_if_empty(expr: LinearExpression) -> bool:
     """Check if the expression is empty.
+
     This is a workaround for the issue that linopy does not support
     the empty property for older versions (`.empty` in >=0.5.1 vs `.empty()` in <0.5.1).
     """
@@ -70,10 +70,10 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
             grouper = by
 
         grouper.insert(0, "component", c)  # for tracking the component
-        return grouper.rename_axis(c)
+        return grouper
 
     def _get_component_index(self, obj: LinearExpression, c: str) -> pd.Index:
-        return obj.indexes[c]
+        return obj.indexes["name"]
 
     def _concat_periods(self, exprs: dict[str, LinearExpression], c: str) -> Any:
         return ln.merge(list(exprs.values()), dim=c)
@@ -150,7 +150,7 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
 
     def _get_operational_variable(self, c: str) -> Variable | LinearExpression:
         # TODO: move function to better place to avoid circular imports
-        from pypsa.optimization.optimize import lookup
+        from pypsa.optimization.optimize import lookup  # noqa: PLC0415
 
         m = self._n.model
 
@@ -188,11 +188,10 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
         def func(n: Network, c: str, port: str) -> pd.Series | None:
             m = n.model
             capacity = m.variables[f"{c}-{nominal_attrs[c]}"]
-            capacity = capacity.rename({f"{c}-ext": c})
             if include_non_extendable:
                 query = f"~{nominal_attrs[c]}_extendable"
                 capacity = capacity + n.df(c).query(query)["p_nom"]
-            costs = n.df(c)[cost_attribute][capacity.indexes[c]]
+            costs = n.df(c)[cost_attribute][capacity.indexes["name"]]
             return capacity * costs
 
         return self._aggregate_components(
@@ -241,11 +240,10 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
             m = n.model
             attr = nominal_attrs[c]
             capacity = m.variables[f"{c}-{nominal_attrs[c]}"]
-            capacity = capacity.rename({f"{c}-ext": c})
             if include_non_extendable:
                 query = f"~{attr}_extendable"
                 capacity = capacity + n.df(c).query(query)[attr]
-            efficiency = port_efficiency(n, c, port=port)[capacity.indexes[c]]
+            efficiency = port_efficiency(n, c, port=port)[capacity.indexes["name"]]
             if not at_port:
                 efficiency = abs(efficiency)
             res = capacity * efficiency
@@ -265,7 +263,7 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
             nice_names=nice_names,
         )
 
-    def opex(
+    def opex(  # noqa: D417
         self,
         comps: str | Sequence[str] | None = None,
         aggregate_time: str | bool = "sum",
@@ -293,7 +291,7 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
             using snapshot weightings. With False the time series is given in currency/hour. Defaults to 'sum'.
 
         """
-        from pypsa.optimization.optimize import lookup
+        from pypsa.optimization.optimize import lookup  # noqa: PLC0415
 
         @pass_none_if_keyerror
         def func(n: Network, c: str, port: str) -> pd.Series | None:
@@ -303,7 +301,7 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
             var = n.model.variables[f"{c}-{attr}"]
             sns = var.indexes["snapshot"]
             opex = var * n.get_switchable_as_dense(c, "marginal_cost").loc[sns]
-            weights = get_weightings(n, c).loc[sns]
+            weights = n.snapshot_weightings.objective.loc[sns]
             return self._aggregate_timeseries(opex, weights, agg=aggregate_time)
 
         return self._aggregate_components(
@@ -318,7 +316,7 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
             nice_names=nice_names,
         )
 
-    def transmission(
+    def transmission(  # noqa: D417
         self,
         comps: Collection[str] | str | None = None,
         aggregate_time: str | bool = "sum",
@@ -330,8 +328,9 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
         carrier: str | Sequence[str] | None = None,
         nice_names: bool | None = None,
     ) -> LinearExpression:
-        """Calculate the transmission of branch components in the network. Units
-        depend on the regarded bus carrier.
+        """Calculate the transmission of branch components in the network.
+
+        Units depend on the regarded bus carrier.
 
         If `bus_carrier` is given, only the flow between buses with
         carrier `bus_carrier` is calculated.
@@ -361,7 +360,7 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
             if isinstance(efficiency, pd.DataFrame):
                 efficiency = efficiency.loc[sns]
             p = var.loc[:, idx] * efficiency[idx]
-            weights = get_weightings(n, c).loc[sns]
+            weights = n.snapshot_weightings.generators.loc[sns]
             return self._aggregate_timeseries(p, weights, agg=aggregate_time)
 
         return self._aggregate_components(
@@ -389,8 +388,9 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
         nice_names: bool | None = None,
         kind: str | None = None,
     ) -> LinearExpression:
-        """Calculate the energy balance of components in the network. Positive
-        values represent a supply and negative a withdrawal. Units depend on
+        """Calculate the energy balance of components in the network.
+
+        Positive values represent a supply and negative a withdrawal. Units depend on
         the regarded bus carrier.
 
         For information on the list of arguments, see the docs in
@@ -426,7 +426,7 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
             if isinstance(efficiency, pd.DataFrame):
                 efficiency = efficiency.loc[sns]
             sign = n.df(c).get("sign", 1.0)
-            weights = get_weightings(n, c).loc[sns]
+            weights = n.snapshot_weightings.generators.loc[sns]
             coeffs = DataArray(efficiency * sign)
             if kind == "supply":
                 coeffs = coeffs.clip(min=0)
@@ -465,8 +465,9 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
         carrier: str | Sequence[str] | None = None,
         nice_names: bool | None = None,
     ) -> LinearExpression:
-        """Calculate the supply of components in the network. Units depend on the
-        regarded bus carrier.
+        """Calculate the supply of components in the network.
+
+        Units depend on the regarded bus carrier.
 
         If `bus_carrier` is given, only the supply to buses with carrier
         `bus_carrier` is calculated.
@@ -501,8 +502,9 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
         carrier: str | Sequence[str] | None = None,
         nice_names: bool | None = None,
     ) -> LinearExpression:
-        """Calculate the withdrawal of components in the network. Units depend on
-        the regarded bus carrier.
+        """Calculate the withdrawal of components in the network.
+
+        Units depend on the regarded bus carrier.
 
         If `bus_carrier` is given, only the withdrawal from buses with
         carrier `bus_carrier` is calculated.
@@ -525,7 +527,7 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
             kind="withdrawal",
         )
 
-    def curtailment(
+    def curtailment(  # noqa: D417
         self,
         comps: str | Sequence[str] | None = None,
         aggregate_time: str | bool = "sum",
@@ -561,17 +563,17 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
         def func(n: Network, c: str, port: str) -> pd.Series:
             attr = nominal_attrs[c]
             capacity = (
-                n.model.variables[f"{c}-{attr}"].rename({f"{c}-ext": c})
+                n.model.variables[f"{c}-{attr}"]
                 + n.df(c).query(f"~{attr}_extendable")[attr]
             )
-            idx = capacity.indexes[c]
+            idx = capacity.indexes["name"]
             operation = self._get_operational_variable(c).loc[:, idx]
             sns = operation.indexes["snapshot"]
             p_max_pu = DataArray(n.get_switchable_as_dense(c, "p_max_pu")[idx]).loc[sns]
             # the following needs to be fixed in linopy, right now constants cannot be used for broadcasting
             # TODO curtailment = capacity * p_max_pu - operation
             curtailment = (capacity - operation / p_max_pu) * p_max_pu
-            weights = get_weightings(n, c).loc[sns]
+            weights = n.snapshot_weightings.generators.loc[sns]
             return self._aggregate_timeseries(curtailment, weights, agg=aggregate_time)
 
         return self._aggregate_components(
@@ -586,7 +588,7 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
             nice_names=nice_names,
         )
 
-    def operation(
+    def operation(  # noqa: D417
         self,
         comps: str | Sequence[str] | None = None,
         aggregate_time: str | bool = "mean",
@@ -619,7 +621,7 @@ class StatisticExpressionsAccessor(AbstractStatisticsAccessor):
         def func(n: Network, c: str, port: str) -> pd.Series:
             operation = self._get_operational_variable(c)
             sns = operation.indexes["snapshot"]
-            weights = get_weightings(n, c).loc[sns]
+            weights = n.snapshot_weightings.generators.loc[sns]
             return self._aggregate_timeseries(operation, weights, agg=aggregate_time)
 
         return self._aggregate_components(