+++ test/test_io.py
@@ -23,6 +23,39 @@ except ImportError:
     excel_installed = False
 
 
+def custom_equals(n1, n2, ignore_attrs=None):
+    """
+    Custom equality check that allows certain attributes to be different.
+    Parameters
+    ----------
+    n1, n2 : pypsa.Network
+        Networks to compare
+    ignore_attrs : list of str, optional
+        List of attribute names that are allowed to be different.
+    """
+    if not ignore_attrs:
+        return n1.equals(n2, log_mode="strict")
+
+    # Copy networks to avoid modifying originals
+    n1 = n1.copy()
+    n2 = n2.copy()
+
+    for attr in ignore_attrs:
+        for net in (n1, n2):
+            obj = net
+            parts = attr.split(".")
+            for part in parts[:-1]:
+                if hasattr(obj, part):
+                    obj = getattr(obj, part)
+                else:
+                    break
+            else:
+                if hasattr(obj, parts[-1]):
+                    setattr(obj, parts[-1], None)
+
+    return n1.equals(n2, log_mode="strict")
+
+
 # TODO classes could be further parametrized
 class TestCSVDir:
     @pytest.mark.parametrize(
@@ -49,12 +82,12 @@ class TestCSVDir:
             {"test": {"test": "test", "test2": "test2"}},
         ],
     )
-    def test_csv_io_quotes(self, scipy_network, tmpdir, meta, quotechar="'"):
+    def test_csv_io_quotes(self, scipy_network, tmpdir, meta):
         fn = tmpdir / "csv_export"
         scipy_network.meta = meta
-        scipy_network.export_to_csv_folder(fn, quotechar=quotechar)
+        scipy_network.export_to_csv_folder(fn, quotechar="'")
         imported = pypsa.Network()
-        imported.import_from_csv_folder(fn, quotechar=quotechar)
+        imported.import_from_csv_folder(fn, quotechar="'")
         assert imported.meta == scipy_network.meta
 
     def test_csv_io_Path(self, scipy_network, tmpdir):
@@ -62,28 +95,28 @@ class TestCSVDir:
         scipy_network.export_to_csv_folder(fn)
         pypsa.Network(fn)
 
-    def test_csv_io_multiindexed(self, ac_dc_network_mi, tmpdir):
+    def test_csv_io_multiindexed(self, ac_dc_periods, tmpdir):
         fn = tmpdir / "csv_export"
-        ac_dc_network_mi.export_to_csv_folder(fn)
+        ac_dc_periods.export_to_csv_folder(fn)
         m = pypsa.Network(fn)
         pd.testing.assert_frame_equal(
             m.generators_t.p,
-            ac_dc_network_mi.generators_t.p,
+            ac_dc_periods.generators_t.p,
         )
 
-    def test_csv_io_shapes(self, ac_dc_network_shapes, tmpdir):
+    def test_csv_io_shapes(self, ac_dc_shapes, tmpdir):
         fn = tmpdir / "csv_export"
-        ac_dc_network_shapes.export_to_csv_folder(fn)
+        ac_dc_shapes.export_to_csv_folder(fn)
         m = pypsa.Network(fn)
         assert_geodataframe_equal(
             m.shapes,
-            ac_dc_network_shapes.shapes,
+            ac_dc_shapes.shapes,
             check_less_precise=True,
         )
 
-    def test_csv_io_shapes_with_missing(self, ac_dc_network_shapes, tmpdir):
+    def test_csv_io_shapes_with_missing(self, ac_dc_shapes, tmpdir):
         fn = tmpdir / "csv_export"
-        n = ac_dc_network_shapes.copy()
+        n = ac_dc_shapes.copy()
         n.shapes.loc["Manchester", "geometry"] = None
         n.export_to_csv_folder(fn)
         m = pypsa.Network(fn)
@@ -97,14 +130,32 @@ class TestCSVDir:
         sys.version_info < (3, 13) or sys.platform not in ["linux", "darwin"],
         reason="Unstable test in CI. Remove with 1.0",
     )
-    def test_io_equality(self, network_all, tmp_path):
+    def test_io_equality(self, networks_including_solved, tmp_path):
         """
         Test if the network is equal after export and import using CSV format.
         """
-        n = network_all
+        n = networks_including_solved
+        if n.has_scenarios:
+            with pytest.raises(
+                NotImplementedError,
+                match="Stochastic networks are not supported*",
+            ):
+                n.export_to_csv_folder(tmp_path / "network")
+            return
         n.export_to_csv_folder(tmp_path / "network")
         n3 = pypsa.Network(tmp_path / "network")
-        assert n.equals(n3, log_mode="strict")
+        # Allow difference for solved networks
+        # TODO: Remove _components.links with #1128
+        ignore = (
+            [
+                "_components.sub_networks.static.obj",
+                "_components.links",
+                "_components.lines",
+            ]
+            if n.model is not None
+            else []
+        )
+        assert custom_equals(n, n3, ignore_attrs=ignore)
 
 
 class TestNetcdf:
@@ -138,34 +189,34 @@ class TestNetcdf:
 
         assert (imported_sns == exported_sns).all()
 
-    def test_netcdf_io_multiindexed(self, ac_dc_network_mi, tmpdir):
+    def test_netcdf_io_multiindexed(self, ac_dc_periods, tmpdir):
         fn = tmpdir / "netcdf_export.nc"
-        ac_dc_network_mi.export_to_netcdf(fn)
+        ac_dc_periods.export_to_netcdf(fn)
         m = pypsa.Network(fn)
         pd.testing.assert_frame_equal(
             m.generators_t.p,
-            ac_dc_network_mi.generators_t.p,
+            ac_dc_periods.generators_t.p,
         )
         pd.testing.assert_frame_equal(
             m.snapshot_weightings,
-            ac_dc_network_mi.snapshot_weightings[
+            ac_dc_periods.snapshot_weightings[
                 m.snapshot_weightings.columns
             ],  # reset order
         )
 
-    def test_netcdf_io_shapes(self, ac_dc_network_shapes, tmpdir):
+    def test_netcdf_io_shapes(self, ac_dc_shapes, tmpdir):
         fn = tmpdir / "netcdf_export.nc"
-        ac_dc_network_shapes.export_to_netcdf(fn)
+        ac_dc_shapes.export_to_netcdf(fn)
         m = pypsa.Network(fn)
         assert_geodataframe_equal(
             m.shapes,
-            ac_dc_network_shapes.shapes,
+            ac_dc_shapes.shapes,
             check_less_precise=True,
         )
 
-    def test_netcdf_io_shapes_with_missing(self, ac_dc_network_shapes, tmpdir):
+    def test_netcdf_io_shapes_with_missing(self, ac_dc_shapes, tmpdir):
         fn = tmpdir / "netcdf_export.nc"
-        n = ac_dc_network_shapes.copy()
+        n = ac_dc_shapes.copy()
         n.shapes.loc["Manchester", "geometry"] = None
         n.export_to_netcdf(fn)
         m = pypsa.Network(fn)
@@ -176,7 +227,7 @@ class TestNetcdf:
         )
 
     def test_netcdf_from_url(self):
-        url = "https://github.com/PyPSA/PyPSA/raw/master/examples/scigrid-de/scigrid-with-load-gen-trafos.nc"
+        url = "https://github.com/PyPSA/PyPSA/raw/master/examples/networks/scigrid-de/scigrid-de.nc"
         pypsa.Network(url)
 
     def test_netcdf_io_no_compression(self, scipy_network, tmpdir):
@@ -220,14 +271,25 @@ class TestNetcdf:
         sys.version_info < (3, 13) or sys.platform not in ["linux", "darwin"],
         reason="Unstable test in CI. Remove with 1.0",
     )
-    def test_io_equality(self, network_all, tmp_path):
+    def test_io_equality(self, networks_including_solved, tmp_path):
         """
         Test if the network is equal after export and import using netCDF format.
         """
-        n = network_all
+        n = networks_including_solved
         n.export_to_netcdf(tmp_path / "network.nc")
         n2 = pypsa.Network(tmp_path / "network.nc")
-        assert n.equals(n2, log_mode="strict")
+        # Allow difference for solved networks
+        # TODO: Remove _components.links with #1128
+        ignore = (
+            [
+                "_components.sub_networks.static.obj",
+                "_components.links",
+                "_components.lines",
+            ]
+            if n.model is not None
+            else []
+        )
+        assert custom_equals(n, n2, ignore_attrs=ignore)
 
 
 @pytest.mark.skipif(not tables_installed, reason="PyTables not installed")
@@ -253,28 +315,28 @@ class TestHDF5:
         scipy_network.export_to_hdf5(fn)
         pypsa.Network(fn)
 
-    def test_hdf5_io_multiindexed(self, ac_dc_network_mi, tmpdir):
+    def test_hdf5_io_multiindexed(self, ac_dc_periods, tmpdir):
         fn = tmpdir / "hdf5_export.h5"
-        ac_dc_network_mi.export_to_hdf5(fn)
+        ac_dc_periods.export_to_hdf5(fn)
         m = pypsa.Network(fn)
         pd.testing.assert_frame_equal(
             m.generators_t.p,
-            ac_dc_network_mi.generators_t.p,
+            ac_dc_periods.generators_t.p,
         )
 
-    def test_hdf5_io_shapes(self, ac_dc_network_shapes, tmpdir):
+    def test_hdf5_io_shapes(self, ac_dc_shapes, tmpdir):
         fn = tmpdir / "hdf5_export.h5"
-        ac_dc_network_shapes.export_to_hdf5(fn)
+        ac_dc_shapes.export_to_hdf5(fn)
         m = pypsa.Network(fn)
         assert_geodataframe_equal(
             m.shapes,
-            ac_dc_network_shapes.shapes,
+            ac_dc_shapes.shapes,
             check_less_precise=True,
         )
 
-    def test_hdf5_io_shapes_with_missing(self, ac_dc_network_shapes, tmpdir):
+    def test_hdf5_io_shapes_with_missing(self, ac_dc_shapes, tmpdir):
         fn = tmpdir / "hdf5_export.h5"
-        n = ac_dc_network_shapes.copy()
+        n = ac_dc_shapes.copy()
         n.shapes.loc["Manchester", "geometry"] = None
         n.export_to_hdf5(fn)
         m = pypsa.Network(fn)
@@ -288,14 +350,32 @@ class TestHDF5:
         sys.version_info < (3, 13) or sys.platform not in ["linux", "darwin"],
         reason="Unstable test in CI. Remove with 1.0",
     )
-    def test_io_equality(self, network_all, tmp_path):
+    def test_io_equality(self, networks_including_solved, tmp_path):
         """
         Test if the network is equal after export and import using HDF5 format.
         """
-        n = network_all
+        n = networks_including_solved
+        if n.has_scenarios:
+            with pytest.raises(
+                NotImplementedError,
+                match="Stochastic networks are not supported*",
+            ):
+                n.export_to_hdf5(tmp_path / "network.h5")
+            return
         n.export_to_hdf5(tmp_path / "network.h5")
         n5 = pypsa.Network(tmp_path / "network.h5")
-        assert n.equals(n5, log_mode="strict")
+        # Allow difference for solved networks
+        # TODO: Remove _components.links with #1128
+        ignore = (
+            [
+                "_components.sub_networks.static.obj",
+                "_components.links",
+                "_components.lines",
+            ]
+            if n.model is not None
+            else []
+        )
+        assert custom_equals(n, n5, ignore_attrs=ignore)
 
 
 @pytest.mark.skipif(not excel_installed, reason="openpyxl not installed")
@@ -329,33 +409,33 @@ class TestExcelIO:
         imported_sns = pypsa.Network(fn).snapshots
         assert (imported_sns == exported_sns).all()
 
-    def test_excel_io_multiindexed(self, ac_dc_network_mi, tmpdir):
+    def test_excel_io_multiindexed(self, ac_dc_periods, tmpdir):
         fn = tmpdir / "excel_export.xlsx"
-        ac_dc_network_mi.export_to_excel(fn)
+        ac_dc_periods.export_to_excel(fn)
         m = pypsa.Network(fn)
         pd.testing.assert_frame_equal(
             m.generators_t.p,
-            ac_dc_network_mi.generators_t.p,
+            ac_dc_periods.generators_t.p,
         )
         pd.testing.assert_frame_equal(
             m.snapshot_weightings,
-            ac_dc_network_mi.snapshot_weightings[m.snapshot_weightings.columns],
+            ac_dc_periods.snapshot_weightings[m.snapshot_weightings.columns],
             check_dtype=False,  # TODO Remove once validation layer leads to safer types
         )
 
-    def test_excel_io_shapes(self, ac_dc_network_shapes, tmpdir):
+    def test_excel_io_shapes(self, ac_dc_shapes, tmpdir):
         fn = tmpdir / "excel_export.xlsx"
-        ac_dc_network_shapes.export_to_excel(fn)
+        ac_dc_shapes.export_to_excel(fn)
         m = pypsa.Network(fn)
         assert_geodataframe_equal(
             m.shapes,
-            ac_dc_network_shapes.shapes,
+            ac_dc_shapes.shapes,
             check_less_precise=True,
         )
 
-    def test_excel_io_shapes_with_missing(self, ac_dc_network_shapes, tmpdir):
+    def test_excel_io_shapes_with_missing(self, ac_dc_shapes, tmpdir):
         fn = tmpdir / "excel_export.xlsx"
-        n = ac_dc_network_shapes.copy()
+        n = ac_dc_shapes.copy()
         n.shapes.loc["Manchester", "geometry"] = None
         n.export_to_excel(fn)
         m = pypsa.Network(fn)
@@ -369,14 +449,32 @@ class TestExcelIO:
         sys.version_info < (3, 13) or sys.platform not in ["linux", "darwin"],
         reason="Unstable test in CI. Remove with 1.0",
     )
-    def test_io_equality(self, network_all, tmp_path):
+    def test_io_equality(self, networks_including_solved, tmp_path):
         """
         Test if the network is equal after export and import using Excel format.
         """
-        n = network_all
+        n = networks_including_solved
+        if n.has_scenarios:
+            with pytest.raises(
+                NotImplementedError,
+                match="Stochastic networks are not supported*",
+            ):
+                n.export_to_excel(tmp_path / "network.xlsx")
+            return
         n.export_to_excel(tmp_path / "network.xlsx")
         n4 = pypsa.Network(tmp_path / "network.xlsx")
-        assert n.equals(n4, log_mode="strict")
+        # Allow difference for solved networks
+        # TODO: Remove _components.links with #1128
+        ignore = (
+            [
+                "_components.sub_networks.static.obj",
+                "_components.links",
+                "_components.lines",
+            ]
+            if n.model is not None
+            else []
+        )
+        assert custom_equals(n, n4, ignore_attrs=ignore)
 
     def test_io_time_dependent_efficiencies_excel(self, tmpdir):
         n = pypsa.Network()
@@ -414,28 +512,43 @@ class TestExcelIO:
     sys.version_info < (3, 13) or sys.platform not in ["linux", "darwin"],
     reason="Unstable test in CI. Remove with 1.0",
 )
-def test_io_equality(network_all, tmp_path):
+def test_io_equality(networks_including_solved, tmp_path):
     """
     Test if the network is equal after export and import.
     """
-    n = network_all
+    n = networks_including_solved
     n.export_to_netcdf(tmp_path / "network.nc")
     n2 = pypsa.Network(tmp_path / "network.nc")
-    assert n.equals(n2, log_mode="strict")
+    # Allow difference for solved networks
+    # TODO: Remove _components.links with #1128
+    ignore = (
+        [
+            "_components.sub_networks.static.obj",
+            "_components.links",
+            "_components.lines",
+        ]
+        if n.model is not None
+        else []
+    )
+    assert custom_equals(n, n2, ignore_attrs=ignore)
+
+    # Only check with supported io formats
+    if n.has_scenarios:
+        return
 
     n.export_to_csv_folder(tmp_path / "network")
     n3 = pypsa.Network(tmp_path / "network")
-    assert n.equals(n3, log_mode="strict")
+    assert custom_equals(n, n3, ignore_attrs=ignore)
 
     if excel_installed:
         n.export_to_excel(tmp_path / "network.xlsx")
         n4 = pypsa.Network(tmp_path / "network.xlsx")
-        assert n.equals(n4, log_mode="strict")
+        assert custom_equals(n, n4, ignore_attrs=ignore)
 
     if tables_installed:
         n.export_to_hdf5(tmp_path / "network.h5")
         n5 = pypsa.Network(tmp_path / "network.h5")
-        assert n.equals(n5, log_mode="strict")
+        assert custom_equals(n, n5, ignore_attrs=ignore)
 
 
 @pytest.mark.skipif(
@@ -573,3 +686,14 @@ def test_sort_attrs():
     empty_df = pd.DataFrame()
     result = _sort_attrs(empty_df, ["a", "b"], axis=1)
     assert result.empty
+
+
+def test_version_warning(caplog):
+    # Assert no info logged with "version"
+    n = pypsa.examples.ac_dc_meshed()
+    assert "Importing network from PyPSA version" not in caplog.text
+
+    n._pypsa_version = "0.10.0"
+    n.export_to_netcdf("test.nc")
+    pypsa.Network("test.nc")
+    assert "Importing network from PyPSA version v0.10.0" in caplog.text